catching up master for 0.9.3

+++ openmdao.main/src/openmdao/main/mp_support.py
@@ -41,12 +41,13 @@ import hashlib
 import inspect
 import logging
 import os
-import signal
 import socket
 import sys
 import threading
 import time
 import traceback
+
+from Crypto import Random
 
 from multiprocessing import Process, current_process, connection, util
 from multiprocessing.forking import Popen
@@ -57,12 +58,12 @@ from multiprocessing.managers import BaseManager, BaseProxy, RebuildProxy, \
 if sys.platform == 'win32':  #pragma no cover
     from _multiprocessing import win32
 
-from enthought.traits.trait_handlers import TraitDictObject
+from traits.trait_handlers import TraitDictObject
 
 from openmdao.main.interfaces import obj_has_interface
 from openmdao.main.mp_util import decrypt, encrypt, is_legal_connection, \
                                   keytype, make_typeid, public_methods, \
-                                  SPECIALS
+                                  tunnel_address, SPECIALS
 from openmdao.main.rbac import AccessController, RoleError, check_role, \
                                need_proxy, Credentials, \
                                get_credentials, set_credentials
@@ -82,24 +83,31 @@ CLASSES_TO_PROXY = []
 _PROXY_CACHE = {}
 
 
-def is_instance(obj, typ):
+def is_instance(obj, type_info):
     """
     :func:`isinstance` replacement for when `obj` might be a proxy.
 
     obj: object
         Object to be tested.
 
-    typ: class
-        Class to be tested against.
+    type_info: class or tuple of classes
+        Class(es) to be tested against.
 
-    Returns True if `obj` is an instance of `typ` or the object `obj` refers
-    to is an instance of `typ`.
+    Returns True if `obj` is an instance of `type_info` or the object `obj`
+    refers to is an instance of `type_info`.
     """
     if isinstance(obj, OpenMDAO_Proxy):
-        typename = '%s.%s' % (typ.__module__, typ.__name__)
-        return obj.__is_instance__(typename)
+        try:
+            type_info[0]
+        except TypeError:
+            type_info = (type_info,)
+        for typ in type_info:
+            typename = '%s.%s' % (typ.__module__, typ.__name__)
+            if obj.__is_instance__(typename):
+                return True
+        return False
     else:
-        return isinstance(obj, typ)
+        return isinstance(obj, type_info)
 
 
 def has_interface(obj, *ifaces):
@@ -149,7 +157,7 @@ class OpenMDAO_Server(Server):
 
     allowed_hosts: list(string)
         Host address patterns to check against.
-        Ignored if `allowed_users` is specified.
+        Optional if `allowed_users` is specified.
 
     allowed_users: dict
         Dictionary of users and corresponding public keys allowed access.
@@ -157,35 +165,46 @@ class OpenMDAO_Server(Server):
         The host portions of user strings are used for address patterns.
 
     allow_tunneling: bool
-        If True, allow connections from 127.0.0.1 (localhost), even if not
-        listed otherwise.
+        If True, allow connections from the local host, even if not listed
+        otherwise.
     """
 
     def __init__(self, registry, address, authkey, serializer, name=None,
                  allowed_hosts=None, allowed_users=None, allow_tunneling=False):
         super(OpenMDAO_Server, self).__init__(registry, address, authkey,
                                               serializer)
-        self.name = name or 'OMS_%d' % os.getpid()
+        self.name = name or ('OMS_%d' % os.getpid())
+        self._logger = logging.getLogger(self.name)
+        self._logger.info('OpenMDAO_Server process %d started, %r',
+                          os.getpid(), keytype(authkey))
+
+        Random.atfork()  # Get our own PRNG.
+
         self.host = socket.gethostname()
         self._allowed_users = allowed_users
         if self._allowed_users is not None:
             hosts = set()
             for user_host in self._allowed_users.keys():
                 user, host = user_host.split('@')
-                hosts.add(socket.gethostbyname(host))
-                if host == socket.gethostname():
-                    hosts.add('127.0.0.1')
+                try:
+                    ip_addr = socket.gethostbyname(host)
+                except socket.gaierror:
+                    self._logger.warning('No address for %r', host)
+                else:
+                    hosts.add(ip_addr)
+                if host == socket.getfqdn() or host == socket.gethostname():
+                    hosts.add('127.0.0.1')  # localhost
+            if allowed_hosts:
+                hosts |= set(allowed_hosts)
             self._allowed_hosts = list(hosts)
         else:
             self._allowed_hosts = allowed_hosts or []
 
         self._allow_tunneling = allow_tunneling
-        if allow_tunneling and '127.0.0.1' not in self._allowed_hosts:
-            self._allowed_hosts.append('127.0.0.1')
+        if allow_tunneling:
+            self._allowed_hosts.append('127.0.0.1')  # localhost
+            self._allowed_hosts.append(socket.gethostbyname(address[0]))
 
-        self._logger = logging.getLogger(name)
-        self._logger.info('OpenMDAO_Server process %d started, %r',
-                          os.getpid(), keytype(authkey))
         if self._allowed_users is None:
             self._logger.warning('    allowed_users: ANY')
         else:
@@ -243,6 +262,15 @@ class OpenMDAO_Server(Server):
                 while not self.stop:
                     try:
                         conn = self.listener.accept()
+                        # Comment-out the line above and use this equivalent
+                        # to debug connectivity issues.
+                        #conn = self.listener._listener.accept()
+                        #self._logger.critical('connection attempt from %r',
+                        #                      self.listener.last_accepted)
+                        #if self.listener._authkey:
+                        #    connection.deliver_challenge(conn, self.listener._authkey)
+                        #    connection.answer_challenge(conn, self.listener._authkey)
+
                     # Hard to cause this to happen.
                     except (OSError, IOError):  #pragma no cover
                         if self.stop:
@@ -458,7 +486,7 @@ class OpenMDAO_Server(Server):
                 except Exception as exc:
                     self._logger.exception('%s %s %s failed:',
                                            methodname, role, credentials)
-                    msg = ('#ERROR', exc)
+                    msg = ('#TRACEBACK', traceback.format_exc())
                 else:
                     msg = self._form_reply(res, ident, methodname, function,
                                            args, access_controller, conn)
@@ -927,28 +955,31 @@ class OpenMDAO_Manager(BaseManager):
 
         writer.close()
         start = time.time()
+        error_msg = None
         for retry in range(timeout):
             if reader.poll(1):
                 break
             if not self._process.is_alive():
-                raise RuntimeError('Server process %d exited: %s'
-                                   % (pid, self._process.exitcode))
+                error_msg = 'Server process %d exited: %s' \
+                            % (pid, self._process.exitcode)
+                break
         # Hard to cause a timeout.
         else:  #pragma no cover
-            et = time.time() - start
-            self._process.terminate()
-            raise RuntimeError('Server process %d startup timed-out in %.2f' \
-                               % (pid, et))
-        error_msg = None
-        try:
-            reply = reader.recv()
-        except Exception as exc:  # str(Exception()) is null, repr() isn't.
-            error_msg = 'Server process %d read failed: %s' \
-                        % (pid, (str(exc) or repr(exc)))
-        else:
-            if isinstance(reply, Exception):
-                error_msg = 'Server process %d startup failed: %s' \
-                             % (pid, (str(reply) or repr(reply)))
+            if error_msg is None:
+                et = time.time() - start
+                self._process.terminate()
+                error_msg = 'Server process %d startup timed-out in %.2f' \
+                            % (pid, et)
+        if error_msg is None:
+            try:
+                reply = reader.recv()
+            except Exception as exc:  # str(Exception()) is null, repr() isn't.
+                error_msg = 'Server process %d read failed: %s' \
+                            % (pid, str(exc) or repr(exc))
+            else:
+                if isinstance(reply, Exception):
+                    error_msg = 'Server process %d startup failed: %s' \
+                                 % (pid, str(reply) or repr(reply))
         if error_msg:
             logging.error(error_msg)
             if cwd:
@@ -1187,6 +1218,17 @@ class OpenMDAO_Proxy(BaseProxy):
         else:
             self._pubkey = self._manager._pubkey
 
+    def _connect(self):
+        """ This version translates tunneled addresses. """
+        util.debug('making connection to manager')
+        name = current_process().name
+        if threading.current_thread().name != 'MainThread':
+            name += '|' + threading.current_thread().name
+        address = tunnel_address(self._token.address)
+        conn = self._Client(address, authkey=self._authkey)
+        dispatch(conn, None, 'accept_connection', (name,))
+        self._tls.connection = conn
+
     def _callmethod(self, methodname, args=None, kwds=None):
         """
         Try to call a method of the referrent and return a copy of the result.
@@ -1298,7 +1340,7 @@ class OpenMDAO_Proxy(BaseProxy):
                 except Exception:
                     pass
             raise RuntimeError(msg)
-        
+
         self._tls.session_key = key_pair.decrypt(server_data[1])
 
     def _incref(self):
@@ -1308,7 +1350,7 @@ class OpenMDAO_Proxy(BaseProxy):
         """
         # Hard to cause this to happen.
         if not OpenMDAO_Proxy.manager_is_alive(self._token.address):  #pragma no cover
-            raise RuntimeError('Cannot connect to manager at %r' 
+            raise RuntimeError('Cannot connect to manager at %r'
                                % (self._token.address,))
 
         conn = _get_connection(self._Client, self._token.address, self._authkey)
@@ -1382,6 +1424,7 @@ class OpenMDAO_Proxy(BaseProxy):
             elif addr_type == 'AF_UNIX':
                 sock = socket.socket(socket.AF_UNIX)
 
+            address = tunnel_address(address)
             try:
                 sock.connect(address)
             except socket.error as exc:
@@ -1530,8 +1573,13 @@ def _auto_proxy(token, serializer, manager=None, authkey=None,
     """
     ProxyType = _make_proxy_type('OpenMDAO_AutoProxy[%s]' % token.typeid,
                                  exposed)
-    proxy = ProxyType(token, serializer, manager=manager, authkey=authkey,
-                      incref=incref, pubkey=pubkey)
+    try:
+        proxy = ProxyType(token, serializer, manager=manager, authkey=authkey,
+                          incref=incref, pubkey=pubkey)
+    except Exception:
+        logging.exception('Auto proxy creation failed for %s at %s:',
+                          token.typeid, token.address)
+        raise
     proxy._isauto = True
     return proxy
 
@@ -1541,6 +1589,7 @@ def _get_connection(_client, address, authkey):
     Get client connection to `address` using `authkey`.
     Avoids dying on 'Interrupted system call'. (Should be in lower layer)
     """
+    address = tunnel_address(address)
     for retry in range(3):
         try:
             conn = _client(address, authkey=authkey)
